% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_kirby21_data.R
\name{delete_kirby21_data}
\alias{delete_kirby21_data}
\title{Delete Kirby21 Imaging Data}
\usage{
delete_kirby21_data(modality = kirby21.base::all_modalities())
}
\arguments{
\item{modality}{modality of images that are to be downloaded.  You
must have the package downloaded for that modality.}
}
\value{
Nothing is returned
}
\description{
This function allows users to remove specific
modalities for Kirby21 data sets.  This allows this package to be 
on CRAN
}
